restore clientHelpers tests
